Refactored Quest Step for extra data
Modified quest step to house some extra data given in the JSON file. (itemReqs).
Works on #25
