+++ src/main/java/objects/quests/QuestStep.java
@@ -3,6 +3,7 @@ package objects.quests;
 import constants.Constant;
 import constants.Skills;
 import objects.Item;
+import objects.requirements.ItemRequirements;
 
 import java.util.ArrayList;
 import java.util.EnumMap;
@@ -10,11 +11,11 @@ import java.util.EnumMap;
 public class QuestStep {
     private int stepNumber;
     private String stepDescription, extraInfo;
-    private ArrayList<String> itemReqs;
+    private ItemRequirements itemReqs;
     private int chunkLocation;
     private boolean completed;
 
-    public QuestStep(int stepNumber, String stepDescription, ArrayList<String> itemReqs, int chunkLocation) {
+    public QuestStep(int stepNumber, String stepDescription, ItemRequirements itemReqs, int chunkLocation) {
         this.stepNumber = stepNumber;
         this.stepDescription = stepDescription;
         this.itemReqs = itemReqs;
@@ -29,19 +30,22 @@ public class QuestStep {
         return stepDescription;
     }
 
-    public ArrayList<String> getItemReqs() {
-        return itemReqs;
-    }
 
     /**
-     *Returns true if section of the quest is completable with current unlocks.
+     * This method is used to check if the section of the quest
+     * is completable with current unlocks.
+     * @param unlockedChunks An Array of all chunk integers the player has access to
+     * @param currentSkills  An EnumMap of the players current in-game stats
+     * @return boolean Returns whether the step is completable with current progress.
      */
     public boolean completable(ArrayList<Integer> unlockedChunks, EnumMap<Skills, Integer> currentSkills){
         //Check Location
         if(!unlockedChunks.contains(chunkLocation)) return false;
 
         //Check Items
-        for (String item:itemReqs) {
+        //Note this does not check quantity (as it assumes once an item is unlocked, it can be gained repeatably)
+        String[] requiredItems = itemReqs.getItems();
+        for (String item:requiredItems) {
             if (Constant.UNLOCKED_ITEM_DATABASE.contains(item)){
                 Item i = Constant.UNLOCKED_ITEM_DATABASE.getElement(item);
                 if(!i.foundIn(unlockedChunks)) return false;
@@ -50,6 +54,9 @@ public class QuestStep {
             }
         }
 
+        /*
+        If later discovered some steps have skill requirements rather than the complete quest, check that here.
+        */
         return true;
     }
